add a --dry-run mode to cleanup-images.py

+++ scripts/cleanup-images.py
@@ -3,9 +3,16 @@
 import os, sys
 import pwd
 
-if os.geteuid():
-    print >> sys.stderr, "Script must be run as root to manipulate permissions"
+dryRun = False
+if len(sys.argv) > 1 and sys.argv[1] == "--dry-run":
+    dryRun = True
+elif len(sys.argv) > 1:
+    print "usage: cleanup-images.py [--dry-run]"
     sys.exit(1)
+
+if not dryRun and not os.geteuid():
+    print >> sys.stderr, "Script not run as root: forcing --dry-run"
+    dryRun = True
 
 from conary import dbstore
 from mint import config
@@ -22,11 +29,21 @@ imageFiles = [x[0] for x in cu.fetchall()]
 
 for baseDir, dirs, files in os.walk(cfg.imagesPath):
     if len(baseDir.split(os.path.sep)) == 6:
-        os.chown(baseDir, isogenUid, apacheGid)
-        os.chmod(baseDir, os.stat(baseDir)[0] & 0777 | 0020)
+        if dryRun:
+            print "chown %d.%d %s" % (isogenUid, apacheGid, baseDir)
+            print "chmod %0o %s" % (os.stat(baseDir)[0] & 0777 | 0020, baseDir)
+        else:
+            os.chown(baseDir, isogenUid, apacheGid)
+            os.chmod(baseDir, os.stat(baseDir)[0] & 0777 | 0020)
         for file in files:
             path = os.path.join(baseDir, file)
             if file not in imageFiles:
-                os.unlink(path)
+                if dryRun:
+                    print "rm %s" % path
+                else:
+                    os.unlink(path)
             else:
-                os.chown(path, isogenUid, apacheGid)
+                if dryRun:
+                    print "chown %d.%d %s" % (isogenUid, apacheGid, path)
+                else:
+                    os.chown(path, isogenUid, apacheGid)
